+++ tests/core/test_parameters.py
@@ -74,3 +74,22 @@ def test_parameter_repr():
     tf.close()
     tf_parameters = load_parameter_file(file_path = tf.name)
     assert_equal(repr(parameters), repr(tf_parameters))
+
+
+def test_parameters_metadata():
+    parameter = tax_benefit_system.parameters.benefits.basic_income
+    assert_equal(parameter.metadata['reference'], 'https://law.gov.example/basic-income/amount')
+    assert_equal(parameter.metadata['unit'], 'currency-EUR')
+    assert_equal(parameter.values_list[0].metadata['reference'], 'https://law.gov.example/basic-income/amount/2015-12')
+    assert_equal(parameter.values_list[0].metadata['unit'], 'currency-EUR')
+    scale = tax_benefit_system.parameters.taxes.social_security_contribution
+    assert_equal(scale.metadata['threshold_unit'], 'currency-EUR')
+    assert_equal(scale.metadata['rate_unit'], '/1')
+
+
+def test_parameter_node_metadata():
+    parameter = tax_benefit_system.parameters.benefits
+    assert_equal(parameter.description, 'Social benefits')
+
+    parameter_2 = tax_benefit_system.parameters.taxes.housing_tax
+    assert_equal(parameter_2.description, 'Housing tax')