+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+### 23.3.0 [#681](https://github.com/openfisca/openfisca-core/pull/681)
+
+* Change the way metadata are declared for Parameter.
+
+Before:
+```YAML
+description: Age of retirement
+reference: https://wikipedia.org/wiki/retirement
+unit: year
+values: (...)
+```
+
+After:
+```YAML
+description: Age of retirement
+metadata:
+  reference: https://wikipedia.org/wiki/retirement
+  unit: year
+values: (...)
+```
+
+_Setting `unit` and `reference` out of `metadata` is considered deprecated, but still works for backward compatibility._
+
+* Allow legislation coders to define their own medatada
+
+* Expose in the Python API
+    - Parameters metadata:
+      - e.g. `parameters.taxes.rate.metadata['unit']`
+    - Parameter value metadata:
+      - e.g. `parameters.taxes.rate.values_list[0].metadata['unit']`
+    - Parameter node description and metadata:
+      - e.g. `parameters.taxes.metadata['reference']`, `parameters.taxes.description`
+      - Note: Parameter descriptions (e.g. `parameters.taxes.rate.description`) were already exposed
+
 ## 23.2.0 [#689](https://github.com/openfisca/openfisca-core/pull/689)
 
 * Introduce `TaxBenefitSystem.replace_variable`